+++ OndselMbD/GESpMatFullPv.cpp
@@ -114,7 +114,7 @@ void GESpMatFullPv::backSubstituteIntoDU()
 	//auto rhsZeroElement = this->rhsZeroElement();
 	for (int i = n - 2; i >= 0; i--)
 	{
-		auto& rowi = matrixA->at(i);
+		auto& rowi = matrixA->at((int)i);
 		sum = 0.0; // rhsZeroElement copy.
 		for (auto const& keyValue : *rowi) {
 			auto jj = keyValue.first;
